pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "serial2mcp"
version = "0.1.0"
description = "基于 MCP 协议的智能串口交互工具"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Niusulong", email = "niusulong@example.com"}
]
dependencies = [
    "pyserial>=3.5",
    "fastmcp>=0.2.0",
    "typing-extensions>=4.0.0",
    "structlog>=22.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "pre-commit>=2.20.0"
]
async-mqtt = ["asyncio-mqtt>=0.13.0"]
data-processing = ["numpy>=1.21.0", "protobuf>=3.20.0"]
monitoring = ["prometheus-client>=0.14.0"]

[project.scripts]
serial2mcp = "serial2mcp.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
skip-string-normalization = true

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src/serial2mcp --cov-report=html --cov-report=term"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true